Add New Filemon and Signal Handling (#120)
* Working signal handler

* Allow lambda in callback

* Threadsafe callback set
