+++ tests_and_snippets/c_and_cpp/wspr_loop/transmit.cpp
@@ -9,6 +9,8 @@
 #include "scheduling.hpp"
 #include "signal_handler.hpp"
 #include "arg_parser.hpp"
+#include "led_handler.hpp"
+#include "shutdown_handler.hpp"
 
 #include <atomic>
 #include <chrono>
@@ -47,7 +49,7 @@ void perform_transmission(ModeType mode, double tx_freq)
         }
     }
 
-    std::string mode_name = (wspr_interval.load()  == WSPR_2) ? "WSPR" : "WSPR-15";
+    std::string mode_name = (wspr_interval.load() == WSPR_2) ? "WSPR" : "WSPR-15";
     llog.logS(INFO, "Transmission started for", mode_name, "at",
               std::fixed, std::setprecision(6), tx_freq / 1e6, "MHz.");
     in_transmission.store(true);
@@ -109,7 +111,8 @@ void transmit_loop()
         {
             int current_minute = std::chrono::duration_cast<std::chrono::minutes>(
                                      now.time_since_epoch())
-                                     .count() % 60;
+                                     .count() %
+                                 60;
             int next_target = ((current_minute / 15) + 1) * 15 % 60;
             next_wakeup += std::chrono::minutes(next_target - current_minute);
             next_wakeup -= std::chrono::seconds(
@@ -127,7 +130,8 @@ void transmit_loop()
         {
             std::unique_lock<std::mutex> lock(transmit_mtx);
 
-            if (cv.wait_until(lock, next_wakeup, [] { return exit_wspr_loop.load() || signal_shutdown.load(); }))
+            if (shutdown_cv.wait_until(lock, next_wakeup, []
+                                       { return exit_wspr_loop.load() || signal_shutdown.load(); }))
             {
                 llog.logS(DEBUG, "Transmit loop wake-up due to shutdown request.");
                 break;
@@ -149,7 +153,7 @@ void transmit_loop()
 
         double tx_freq = center_freq_set[freq_index];
 
-        // **PERFORM TRANSMISSION HERE**
+        // ** TODO: PERFORM TRANSMISSION HERE**
         if (tx_freq != 0.0)
         {
             llog.logS(INFO, "Transmitting on frequency:", std::fixed, std::setprecision(6), tx_freq / 1e6, "MHz.");
@@ -160,6 +164,10 @@ void transmit_loop()
             llog.logS(INFO, "Skipping transmission per 0.0 freq request in list.");
         }
 
+        // Defered load if INI changes
+        apply_deferred_changes();
+
+        // Skip to next freq in list.
         freq_index = (freq_index + 1) % center_freq_set.size();
 
         // Check if we've reached the transmission limit
@@ -172,7 +180,7 @@ void transmit_loop()
             {
                 llog.logS(INFO, "Completed all scheduled transmissions. Signaling shutdown.");
                 exit_wspr_loop.store(true); // Set exit flag
-                cv.notify_all();            // Wake up waiting threads
+                shutdown_cv.notify_all();   // Wake up waiting threads
                 break;
             }
         }
